ops/benchmarks: lock mysql version in at 8.0.27
Coming back to benchmarking, it looks like in the latest version of
MySql, Oracle attempts to purge data in the performance schema database.
This was resulting in some interesting errors in our builds, so for now,
we can lock things in @ 8.0.27.

After checking with Justin, it sounds like we don't really need the
performance_schema database to be active during tests, so it makes sense
to explicitly disable it I think. That change is also included in this
CL.

Change-Id: If19ac2e9e086f74722fe8dda0ae20152fcf337aa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1697
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>