persistent_state: Tolerate mixed byte-length keys
Given how datatype comparison works, it's possible that a key k1
compares greater to some key k2, even if when serialized it's actually
shorter.  When checking the prefix of a key to see if we've hit either
end of a range during iteration, previously we'd panic when trying to
get the sub-slice of the key if it ended up being shorter - now, we do a
`min` on the length of the key to make sure that doesn't happen.

Change-Id: Ie21cd6e65343ff6da4d74404f9633a4e784c5ea0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1700
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>