noria-data: Add a DataType::Max variant
For queries that mix equality and either less-than or
less-than-or-equal-to keys, we want to turn the actual lookup key into a
compound range where the equal keys are ordered first followed by the
less-than keys, and then the actual lookup key becomes an inclusive
range where the value for the equal-keys are the same on both the lower
and upper bounds, and the less-than keys have the value on the lower
bound - but we need something to put in the place of the less -than keys
on the upper bound. For types like integers, this can trivially just be
eg `i32::MAX`, but we don't currently have a convenient "max" sentinel
value for unbounded types like strings (we *do* have a convenient "min"
sentinel value for everything, since DataType::None is the minimum
DataType).

This commit adds a new `DataType::Max` variant, which by virtue of being
the last enum variant in in the list compares greater than all other
DataTypes, to be used as the sentinel "max" value in these cases.
Generally, this MAX value has the same semantics as `DataType::None`,
except where proper injectivity matters like in serialization, hashing,
and equality comparisons.

Change-Id: Ibd80db76e231b10b227c3c528f52bfc023fed3b2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1701
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>