consul: Replace `consul` dependency with `consulrs`
This CL performs a significant refactor of the consul authority to
replace our forked `consul` dependency with a (temporarily) forked
`consulrs` dependency.

The `consulrs` crate is much more maintained, cleaner written, includes
proper error handling and implements the key value and session APIs.
It is a less *mature* crate (initial commit Sept 22), while `consul`
dates back years, but as this is a simple code base, wrapping HTTP
requests to Consul I think there is low risk to the change.

We are temporarily forking off of our own branch to prevent using a
dependency for TLS that has a copyleft LICENSE. The author of `consulrs`
has already accepted two PRs to remove an unneccessary 'static and
put the copyleft LICENSE dependency behind a feature flag in a
dependency of `consulrs`, [1][2]. In the event we need to maintain the
fork, this code base is *much* more friendly to maintain than `consul`.

Once [3] is merged we can move consulrs off of our fork.

[1] https://github.com/jmgilman/consulrs/pull/15
[2] https://github.com/jmgilman/rustify/pull/7
[3] https://github.com/jmgilman/consulrs/pull/16

Change-Id: Iaf8e815729416f6b33c71581a5003e3fd4bae6ef
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1712
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>