persistent state: Drop writes with an old replication offset.
If for any reason there are two replicators running in the system we
may have duplicated writes. Add a small check to the persistent state
code to drop writes with replication offsets <= our current replication
offset as we have already applied these writes.

This may occur during leader failure, when a new leader is elected and
the old leader has not yet realized it has lost leadership.

Change-Id: I03fd240a2ddf758f3bac900850680ebb73b61b90
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1713
Reviewed-by: Corbin Hall <sif@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Tested-by: Buildkite CI