docs: Cargo.lock and updating git dependencies via `cargo update`
This adds documentation on what `Cargo.lock` does, how we use it in CI,
when it should be updated. This also adds a section on how to update git
dependencies that depend on a branch with `cargo update`, as the
revision used is locked in `Cargo.lock` at the time the dependency is
added.

Change-Id: Iea8c0592eddeca54c9994485313c06819985639e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1717
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>