Dockerfile: Set installed toolchain as the default
To ensure that all cargo commands run with the installed toolchain
instead of with the latest nightly, use
`rustup default` instead of `rustup toolchain install`, if the toolchain
does not exist, `rustup default` also installs it.

This will allow us to build and work with the version of nightly that we
choose, the latest nightly may have internal compiler errors or bugs.

Change-Id: I2712609702e0f737792f5c274ada73346b3276e0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1720
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>