nom-sql: Parse boolean literals
Both mysql[0] and postgresql[1] accept case-insensitive `TRUE` or
`FALSE` as literals for boolean values - in mysql, these just work the
way we've always been treating booleans, by evaluating as either the
integers 1 or 0, whereas in postgresql there's an actual boolean *type*
which behaves slightly differently. Getting *correct* support for the
latter is going to happen alongside the introduction of type information
to dataflow, but in the meantime this at least gets us proper boolean
literal parsing in mysql, which is needed for the Solidus demo among
other things.

[0]: https://dev.mysql.com/doc/refman/8.0/en/boolean-literals.html
[1]: https://www.postgresql.org/docs/14/datatype-boolean.html

Change-Id: I0eb6c485d536bca200ffea6d680038a106edd5e1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1721
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>