dataflow: Get rid of slice_to_2_tuple
As we long suspected, this function *is* indeed incredibly unsound, and
appears under certain circumstances to be able to lead to a double-drop
on the underlying DataType, which can have a number of nasal-demon[0]
style effects - in my case, the second drop underflowed the reference
count on an underlying `ThinArc` for a `DataType::Text`, causing it to
trigger a process::abort when trying to then *increment* past
usize::MAX. I also saw a segfault on the same workload, but was unable
to fully diagnose where that was happening - I can only assume it was
another symptom of the same underlying issue.

The *actual* cost of these 2 clones, while not zero, is likely going to
be pretty negligible, since any values larger than 16 bytes are stored
behind an Arc.

[0]: http://www.catb.org/jargon/html/N/nasal-demons.html

Change-Id: Ie35df537d0bbbc1023d4f7b0944fc8a1d453eeda
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1722
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>