view: Support NotEqual as a ViewQueryOperator
In addition to LIKE and ILIKE, support NotEqual as a ViewQueryOperator,
passed as part of the ViewQuery to be applied *after* doing the lookup
into the reader map.

This is the current way I'm planning on implementing exclusive
comparisons (`>` and `<`) that are mixed with inclusive
comparisons (`=`, `>=`, and `<=`) in the same query - just do the
inclusive version, then filter out the endpoint. It's a bit of a hack,
and we pay for it slightly (by overquerying a little bit and having to
do extra filtering and iteration in the lookup), but is *quite*
straightforward to implement compared to introducing some sort of
mixed-clusivity range endpoint thingy inside of KeyComparison, which is
the only other way I can think of to do that. If this ends up being
unworkable (or the pathological case where the equal-key we're filtering
out has a *ton* of rows ends up being common enough to care about) we
can revisit making ranges more complex, but for now this ought to be
Good Enough.

Change-Id: Ie06e3328e7985f2eacfbc7daaff6a6db84739622
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1726
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>