server: Add some more tests for QueryGraph::view_key
This started out as an attempt to add support for mixing different
operators, and mixing operators and BETWEEN, but it turned out that all
of that already worked! So at the end of the day, this is just a commit
to add a bunch more tests for this function, proving that it already
supports these things correctly.

Change-Id: Ia6ea9007fd026de7cc3c48073707b8ed351adfe3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1727
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>