ops/consul: disable update checks of server in airgapped envs
For some reason, Consul decides to try to phone home to update itself
when the Consul service starts.

For our purposes, this functionality should be disabled, and so this CL
aims to accomplish that. Reaching out to the internet to do an update,
and "airgapping" don't play well together. Not sure why they prevent the
service from starting in this case, but I digress.

Change-Id: If7709c4e0258f18ce99b6dca9857f079e3a903eb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1728
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI