deny.toml: Skip checking dependencies on out-of-date required crates.
The skip-tree array now includes the set of crates that are unlikely to
have their dependencies updated without manual intervention. Submitting
PRs against these crates will be required to have their dependencies
updated in a timely manner.

This also adds the rustify git repository to the allowed git repos as we
are currently using it in one of our forks and is safe.

Change-Id: I0cba62efbd9d96efb18ba30203d1ec4b3535a98f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1734
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>