noria-client: Factor out ViewQuery construction
Factor out the construction of the ViewQuery for actually performing a
query into its own (pure) function, rather than doing it inline in
do_read. This is going to get a fair bit of additional complexity to
support mixed comparisons (it was already pretty complex to support
post-read [I]LIKE), and I'd like to be able to write unit tests for it.

Change-Id: Ib6e53961c84e477a4fd9d5e227ed2ac6913944a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1737
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>