noria-client: Add support for inclusive mixed comparisons
Add support in build_view_query for making lookup keys for queries that
mix equality and inclusive comparisons, by first checking to see if we
have mixed comparisons at all, and if we do, iteratively constructing
both ends of the range in the same way we do for BETWEEN based on the
comparison operator of the lookup. This all is covered with a new
logictest, which required enabling the allow_mixed_comparisons flag to
get it to run.

Refs: ENG-935
Change-Id: Ifbc39baef3ce7deecbe2a941907f8076fb77e5f5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1739
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>