noria-client: Add support for exclusive mixed comparisons
As foreshadowed in change Ie06e332, implement *exclusive* comparisons
that are mixed with other comparisons by building an inclusive range,
then adding a post-lookup ViewQueryFilter to filter out the endpoint of
the range to make it effectively exclusive. A more thorough discussion
of the tradeoffs with this approach is there, but in summary this is
primarily an expedience, as introducing mixed-clusivity key comparisons
is likely to be a *very* significant change, and there's a fair chance
that this approach is Good Enough for now.

Fixes: ENG-935
Change-Id: I98be4cc85c1739a7a62b454f21b44d623c33fddd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1740
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>