clustertest: Timeout migrations after 1 second.
This makes failures get detected much quicker in clustertests. The
timeout for migration requests is incredibly high so that in production
with large query graphs, we don't timeout for a client request. However,
the graphs we use in clustertests are small and requests are executed
locally. This reduces the timeout to 1 second.

Change-Id: I51a128756913a3592475fdfb40f30f8d86581ab9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1748
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>