dataflow: Add aggregate support to PostLookup
Add the ability for reader::PostLookup to specify a set of aggregate
functions to apply *after* doing the lookup into the reader, to be used
for range queries where the actual lookup range can cover multiple keys
in an operator that otherwise should be grouped.

Since previously we were only shuffling around *existing* DataTypes, but
not ever actually creating *new* values, but now we *are* creating new
DataTypes that exist as the result of the re-aggregation, a bunch of
types needed to change from &DataType to Cow<DataType>, so that those
newly created datatypes can live in Cow::Owned.

Note that this commit is just adding support in the PostLookup struct,
and nothing is filling it in during migration yet.

Change-Id: Ib77ab8c18d71f8dcfa8c1fa3fae89e4b732d8737
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1752
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>