server: Allow *any* duplicated op on the same col
The issue fixed by 9eabd5972 (server: Don't error out on double-equality
of the same col, 2022-01-24) is actually more general - *any* instance
of doing the same comparison operator on the same column should be
allowed. Similar to that commit, this happens sometimes in generated
logictests, and used to be allowed and should continue to be.

Change-Id: I8700edf40a8ed97616dabeb643f610b702399f05
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1755
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>