leader: Handle all the workers before graph recovery and commit.
Performing a graph recovery and committing the dataflow state should be
done after adding all the new workers to the state. Not including these
workers means: the recovery will not schedule domains on to those
workers, and we will pay the cost of commiting the dataflow state for
every other worker in the set.

This only matters in a world where quorum is not equal to the number of
servers in the system.

Change-Id: I061949f4757e9bfde499c268b647c2b6ebf35902
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1761
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>