noria: Replaced `expected` call with error handling.
There were issues with logictest "project_literal,join_subquery,range_param".
The test attempts to issue a SELECT without a table, which is not supported.
However, instead of returning an error result, we were calling `.expect`, which
was causing the migrations thread to panic and exit.

This commit simply replaces the `.expect` call with the `?`, so if the
`nodes_for_named_query` fails, the error bubbles up.

Change-Id: Ie358d7a648eebe12aaee3c8176e933edf02a20a4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1765
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>