ci: Remove external AMIs pipeline
Currently, the external AMIs pipline means that in order to release to
public we need to run the Rust binary build twice and the Packer build
twice.

Instead, we now release an alpha based upon the git commit always when
the steps are unblocked and then we can also on main builds release
from that build.

Change-Id: Idbb5d707390ee8ca2b06d5377f2d49ffb00d94ce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1769
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>