server: Thread through aggregates to post-lookup
When building a MIR Leaf node for a query that combines aggregates and
range lookups, carry through information about those aggregates in a new
`aggregates` field on the MirNodeInner::Leaf variant, then use that
field to specify the `aggregates` field on the `PostLookup` struct
introduced in 92f290491 (dataflow: Add aggregate support to PostLookup,
2022-01-26).

To avoid duplicating the data structures between PostLookup and MIR,
this makes the PostLookupAggregates struct and all its associated
sub-types parametrized on the type of column reference it uses - this
allows us to use the name-based `mir::Column` in MIR but still use
`usize` for column indices in dataflow, just like we do everywhere else.
The mapping from mir::Column to usize in these structs is done via a new
`map_columns` method.

Fixes: ENG-937
Change-Id: I4e3af61048dc30d77688b3eb1f270dd2eb7c2e3f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1777
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>