Implement eviction strategies for readers
This change adds an option to use different eviction
strategies for the reader nodes.

The change is rather simple. One part is adding a shared EvictionMeta
to both parts of the left_right map, and make sure that it is properly
applied to both the left and the right sides across all operations.
The other part is to modify the meta on each read according to the
chosen eviction strategy, and then evict accordingly when asked by
backlog.

Change-Id: I492d9f6e4d868f048811a1c3e42bb790faae6e9b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1775
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>