ops/gerrit: Don't trigger builds on merged CLs
The gerrit ref-updated hook gets called for CLs when they're merged,
in *addition* to for the updated refs/heads/main ref - building both is
just a waste of time and resources, and has caused inflated queue times.

This change just checks before actually triggering the build for a CL if
that CL is an ancestor of `main`, and if so exits early with a message.

Change-Id: I53a5c46decfee7f0db50f6db9eb6182be267036f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1784
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>