benchmarks: Support running noria-server and noria-mysql in process.
This adds two flags `--local` and `--local-with-mysql` that both run
noria-server and noria-mysql in process on the same tokio backend as the
benchmarks. `--local-with-mysql` can be passed a MySQL connection string
that may be used as an upstream database.

Some of the benchmark schema files include SET statements that Noria
alone does not handle, these will work with `--local-with-mysql`.

Change-Id: I36c530caa343a374534c4218273829abd9a982f8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1788
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>