benchmarks: Metrics labels for benchmarks results and append results
This CL requires that benchmark labels be passed with their respective
metrics::Unit. This allows us to render units when outputting benchmark
results. The existing set of benchmarks are updated to include these new
units.

This makes the results-file only include human readable results, instead
of serializing the YAML as the YAML of the struct is much less readable.
In the future we will be introducing comparing results, which will
require serialization of structs, but this will be separate from the
human readable results file.

Additionally this removes overwriting the results file, instead results
are appended to the file if it exists, and is created otherwise.

Change-Id: I5dfb7991b03520bf264b59153530b56dab0542cf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1785
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>