benchmarks: Better parameter failure handling and warnings.
This adds parameter checking for if the deployment and benchmarks are
passed through some acceptable method, otherwise an error is returned.
Additionally, if users pass parameters that would be overwritten by
another parameter, this will now log a warning.

These fill the gaps where it is not possible to easily rely on clap
attributes. Largely due to the functionality that we can pull in a
entire flattened struct through YAML.

Change-Id: Ib9dbb946202bd964e9c364626102102f5a738c10
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1789
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>