benchmarks: More robust paths for specifying benchmarks/YAML
This introduces the `benchmark_path` utility function which allows
searching for the specified benchmark in the CARGO_MANIFEST_DIR if it
is a relative path and it can't be found from the current working
directory. It will also search the current directory recursively if all
else fails. This only considers directories in the walk, so should not
be too expensive even when called from the ReadySet root.

The YAML files are also updated to use relative paths that only include
the benchmark's immediate subdirectory representing the benchmark's
group and the benchmark name, i.e. irl/query3.sql rather than
src/data/irl/query3.sql.

Change-Id: Ic26bddd048ae6831b1495476256fad3500d87e91
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1790
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>