flakiness: Replace flaky_finder with detect_flake.
flaky_finder has an AGPL license. It *seems* we are safe to use it as
an internal tool, however, since we include it in our Dockerfile's I'm
not too sure the implications. It is safer to replace it with an MIT
licensed flakiness tool that way we can also make modifications if
necessary without AGPL licensing the modifications.

Change-Id: I3185536d09ad40d0ee805fea5bde7e67f1c24f7b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1791
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>