benchmarks: Don't recreate database on --local.
This provides a quality of life changes to benchmarks. We do
not recreate the database when --local-with-mysql is used, as this makes
undesirable behavior occur when used with --skip-setup: the tables don't
exist in the upstream database as expected. As we now may recreate the
noria deployment every iteration, we add waiting for snapshotting during
each iteration.

Change-Id: I81c8ec1a6447ef363bebe4672ceb99bafc158909
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1792
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>