ops/benchmarks: add complete benchmarking documentation & refactor a bit
Adding a section for benchmarking documentation in our docs/ directory.

Have this compartmentalization should help us maintain the complexity of
the benchmark docs as the docs grow over time.

Adjusted some headings slightly so that the heading structure displays
as I think we were expecting.

Fixed a minor type in a MD page I was using as a reference :)

Change-Id: I405dd32f7a44d3e7e2e2dfeb720b83c456d68b27
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1793
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>