clustertest: Add leader failover benchmark
This adds a benchmark to clustertest which measures the time it takes
for cluster leadership to fail over to a new server. As the benchmark
is fairly slow (~40s per iteration on my laptop, >6m for the shortest
run criterion supports, >1h for default), it's hidden behind the
"slow_bench" feature flag to keep it off by defaut.

Change-Id: If2733133a534deb33b93a19c950bf9fd78d24ce1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1795
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI