noria: Refactored `Recipe` and improved recovery mechanism.
This commit refactors the `Recipe` to be only in charge of handling the
MIR graph and processing any incoming changes.
The actual changes (and the parsing of the SQL queries) was delegated to
 a new module (`changelist.rs`).

Having a list of changes improves the readability (and I might say,
performance) of migrations greatly, as we previously had to:
1. Clone the existing recipe
2. Make changes to the cloned recipe
3. Apply changes (which was done by calculating the diff between the old
 recipe and the cloned one).
With these changes, we do not need to do that anymore.

These changes also required that we change (improve) the way we handle
recovery in the event some workers failed.
Previously, the algorithm went like this (this is a simplification):
1. Gather all the nodes that belonged to the failed workers.
2. Remove the workers from the `Leader`.
3. Clone our recipe, and remove those nodes from the clone.
4. Apply the clone recipe, but keeping a copy of the old recipe.
5. Apply recipe again. This triggered the same logic to re-add the
nodes, but without the possibility of using the same workers again.
The logic is now reduced to this:
1. Gather all the domains and their nodes that belonged to the failed
workers.
2. Remove the workers from the `Leader`.
3. Use the recover mechanism we already have, but to recover only the
affected domains and their nodes.

The downside of this commit is that some of the tests that we previously
 had in the `Recipe` do not make sense anymore:
1. The tests on parsing were salvaged and moved to the `changelist.rs`
module.
2. There are some tests that had to be deleted, since changing them
would require us testing `Recipe::activate`, which is hard to do with
the current state of the system (where we require passing a `Migration`
to the `Recipe::activate` method).
Ideally, we should remove the `Migration` from the method, and instead
return the list changes to be added to a `Migration`, to finally
decouple the `Recipe` from it.

Refs: ENG-92
Change-Id: I3da5099cc276caa4c7c6bb2121a01ffee56123c3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1802
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>