noria: Basic `DROP TABLE` support.
We already had support for removing base tables. Here we are extending
those capabilities.
- In MIR: When we remove a base table, we also make sure to remove any
MIR query that might have the base query as its root.
- In dataflow graph: We remove all the children from the base node,
recursively.

There are also a couple of useful tests for `DROP TABLES`.
This commit is only missing the deletion of the persistent state, which
will come in a future commit.

Refs: ENG-92
Change-Id: I202d5fb6af1a966270225efe25d65af6bf71ad04
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1804
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>