persistent_state: Don't drop duplicate rows in upper bound
For a range query into persistent state, there might be *multiple* rows
where the key is equal to the inclusive upper bound of the range. We
want that range query to return all those rows, but previously we'd stop
iterating as soon as we hit the first one. This fixes that, by tracking
whether we've already hit the inclusive upper bound of a range (if any;
note that exclusive upper bounds are already handled by rocksdb) and
stopping iteration only after we *no longer* are pointing to one of
those keys (or if the iterator hits the end of the table, obviously).

Change-Id: I25dcdf3c1732ff96f4cf1c196d8ca17ef6873a43
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1805
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>