Implement compaction progress report
To get any updated stats for the compaction progress, we have
to manyally tap into the LOG file, and extract information
about the creation and indexing of new tables. It seems to
work just fine, however the thing is fragile, and can break
if log format changes in the future. Sadly none of the properties
we could use to get any information about the compaction
progress, such as `rocksdb.num-files-at-level<N>` is updated
during the compaction proccess, so they can't be used for
monitoring, hence this hack.

Change-Id: I891075955313449659081c57497b039c3181db44
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1809
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>