noria: Remove rocksdb persistent state when removing base tables.
This introduces a small refactor as to how we handle the materialized
state:
1. We don't rely on `Box<dyn State>` anymore. Instead, we have a new,
very simple enum called `MaterializedNodeState`, which has the two state
 variants we currently have: `Memory` and `Persistent` (which have
 `MemoryState` and `PersistentState`, respectively).
This new struct also implements `State`, which internally just does a
match and then calls the real method depending on the variant.
2. Thanks to having the enum, we can now have a method called
`tear_down` in the `State`, which takes `mut self`.
This method will be the one that (for the persistent state) takes the
path to the rocksdb directory, drops the rocksdb handle, and removes the
directory, recursively.

This also adds a test to make sure we are deleting the rocksdb data.

Refs: ENG-92
Change-Id: I3d7f96732df1e2730eae9550be628158726dc3c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1820
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>