memory_state: Allow satisfying lookups with shuffled indexes
Similar to how we currently behave for indexes that are a
strict *subset* of other indexes, allow lookups into MemoryState to be
satisfied by indexes that have the same set of columns, but in a
different order.

This fixes an issue similar to, but slightly distinct
from the issue that f4ba18c03 (dataflow: Use tag to select index for
self-lookups, 2022-02-01) fixed - the unifying trait here being that a
node was doing a self-lookup during processing that was different from
the key that was being replayed. In that case, the issue was that the
node was filling a range but doing a self-lookup of a point key, but in
this case the issue is that the replay key and the self-lookup key have
the same set of columns, but in a different order. This caused the same
"do another replay when already replaying" issue, only in this case the
eventual index got *both* the original replay and the second replay
filled, meaning the eventual result was doubled.

This fixes (works around?) the issue by avoiding missing while replaying
in the first place, though there are likely *many* other ways of fixing
the issue that we may want to do either in addition to this or instead
of this.

Similar to the existing method of satisfying misses with other indices,
this is asymptotically quite poor, and we may want to revisit it at some
point in the future - but for now the impact is limited since it only
happens during misses, and the coefficients in general are generally
quite small (nodes usually only have one or two indexes on them anyway)

Change-Id: I32f6b2d81980e40ca5a69aacc1a3bea47e6723d4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1826
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>