ytt-minus: Support example value for replacement
One issue that has appeared with working with the ytt templates is that
the files are not valid YAML because the values are null and so cfn-lint
does not run locally in editors against them.

This updates ytt-minus so that it overrides the value so that locally
the templates still pass cfn-lint and the example value is replaced only
at compile time.

Change-Id: Ie82e4d59bf820e2f9cdc713a33c88ab990f5691c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1833
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>