reader: Special-case reversed ranges to return no results
`BTreeMap::lookup_range` panics if it's passed a range where the upper
bound is less than the lower bound - this is something that *could*
hypothetically happen if a user does a BETWEEN-style lookup with this
kind of range, in which case it's better to just return an empty result
set directly rather than panicking. This adds a case in the reader when
doing the *initial* lookup to check for these kinds of ranges and
short-circuit to return an empty result set immediately.

Change-Id: Iab83b30a6800a206d8a28a8846a91d5bbb88aed4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1837
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>