ops/terraform: Build module to deploy EKS cluster with deployments
The purpose of this CL is to get the base ReadySet EKS cluster Terraform
module rolling for us.

In essence, this module's responsibility is to get the base cluster set
up, as well as to provision all the foundational components that make a
Kubernetes cluster so dynamic and powerful. Follow up CLs will add
things like the LB-Ingress-Controller, Cluster-Autoscaler, etc.

This module hasn't been fully tested yet, so follow up CLs will be used
to refine things as we actually deploy what's proposed here :)

Change-Id: I414d7df8f85c5116ac215e80c6e0e75ab99a45dc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1839
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>