ops/terraform: Implement cluster-autoscaler submodule for eks tf module
The purpose of this TF module is to encapsulate the work of provisioning
IAM resources and the Helm chart for Cluster Autoscaler.

In our infrastructure as code, this module will be called as part of the
overarching "EKS" module. As we deploy new clusters in the future,
foundational components such as cluster-autoscaler can be deployed
automatically without additional work on our part.

Change-Id: Id95ac7dec81f5d63963597dfb622cc41ce6f3292
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1840
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>