ops/terraform: add tf module to tag vpc subnets for k8s auto-discovery
Certain k8s charts such as the LB Ingress Controller help reduce the
amount of manual configuration that's needed on behalf of operator by
implementing auto-discovery of public and private subnets within your
vpc.

The way these deployments accomplish this is by reading certain tags
from your VPC's subnets, and automatically registering them as either
private or public, and determining whether an EKS cluster lives within
the VPC to begin with. This

The goal of this CL is to accomplish adding those tags retroactively to
a target VPC so that auto-discovery mechanisms may work as expected.
We'll see if AWS RAM + Substrate prevent this approach from working, but
we can play by ear.

Change-Id: I7863f6ce860c39564366a1efa448d960931f325b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1841
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>