ops/terraform: add node-termination-handler TF module for graceful nterm
This is another CL to add a submodule to the EKS module. The purpose
here is to implement a mechanism for AWS to gracefully drain pods off
EC2 instances which are being terminated. Having this functionality is
critical towards being able to treat servers as ephemeral without
causing service disruptions.

This module hasn't been fully tested yet, but will be once we
instantiate the overarching eks module in one of our root-modules.

Change-Id: I48d6b76a8ab03486bd2d2504475bd1d7b02c9448
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1842
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>