ops/terraform: add External-DNS submodule for dynamic DNS records
Adding a submodule that will be deployed as part of the eks terraform
module.

The purpose of externalDns is to allow one to specify hostname for a
particular ingress resource in Kubernetes, and ExternalDNS will
automatically create DNS records that map to that ingress.

Having ExtDNS makes it so we don't need to go around manually creating
new DNS records each and every time we deploy an app to a new
environment, etc.

ExternalDNS is capable of creating records in both private and public
hosted zones in R53, and so this module is built to support both models.
This It is worth noting that separating internal/external ExternalDNS (I
know the name is confusing there) into separate deployments can save
some headaches down the road, so it's suggested to separate them into
bespoke deployments. This CL aims to support that idea.

Change-Id: I21ee6ab146f8a54704068e8af8e0801a05e8466e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1843
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>