Improved handling of the TimestampTz DataType
This change inlines the TimestampTz into DataType by
taking advantage of the fact that the offset required
by TimestampTz is at most 18 bits in practice, this
allows us to avoid using the extra overhead and indirection
of an Arc. In addition it improves the serde implementation
byt using a technique similar to the one I already used
in Timestamp. As a result TimestampTz is much faster now

Serde/Serialize TimestampTz
                        time:   [13.608 ns 13.745 ns 13.896 ns]
                        change: [-97.332% -97.302% -97.272%] (p = 0.00 < 0.05)
                        Performance has improved.
Serde/Deserialize TimestampTz
                        time:   [31.034 ns 31.229 ns 31.457 ns]
                        change: [-88.615% -88.473% -88.348%] (p = 0.00 < 0.05)
                        Performance has improved.

Change-Id: I54be08bf61d0c80f5cc7b6ae0090206292578980
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1848
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI